Apply suggestions to merge k8s files
