+++ pkg/client/k8s.go
@@ -24,6 +24,10 @@ import (
 	"path/filepath"
 	"strings"
 
+	appsv1 "k8s.io/api/apps/v1"
+	corev1 "k8s.io/api/core/v1"
+	v1 "k8s.io/api/storage/v1"
+
 	lvmclient "github.com/openebs/lvm-localpv/pkg/generated/clientset/internalclientset"
 	"github.com/openebs/openebsctl/pkg/util"
 	zfsclient "github.com/openebs/zfs-localpv/pkg/generated/clientset/internalclientset"
@@ -238,4 +242,139 @@ func (k K8sClient) GetVersionMapOfComponents() (map[string]string, error) {
 	}
 
 	return versionMap, nil
+}
+
+/**
+CORE RESOURCES
+*/
+// GetPods returns the corev1 Pods based on the label and field selectors
+func (k K8sClient) GetPods(labelSelector string, fieldSelector string, namespace string) (*corev1.PodList, error) {
+	pods, err := k.K8sCS.CoreV1().Pods(namespace).List(context.TODO(), metav1.ListOptions{LabelSelector: labelSelector, FieldSelector: fieldSelector})
+	if err != nil {
+		return nil, fmt.Errorf("error getting pods : %v", err)
+	}
+	return pods, nil
+}
+
+// GetDeploymentList returns the deployment-list with a specific
+// label selector query
+func (k K8sClient) GetDeploymentList(labelSelector string) (*appsv1.DeploymentList, error) {
+	if pv, err := k.K8sCS.AppsV1().Deployments("").List(context.TODO(), metav1.ListOptions{
+		LabelSelector: labelSelector,
+	}); err == nil && len(pv.Items) >= 1 {
+		return pv, nil
+	} else {
+		return nil, fmt.Errorf("got 0 deployments with label-Selector as %s", labelSelector)
+	}
+}
+
+// GetSC returns a StorageClass object using the scName passed.
+func (k K8sClient) GetSC(scName string) (*v1.StorageClass, error) {
+	sc, err := k.K8sCS.StorageV1().StorageClasses().Get(context.TODO(), scName, metav1.GetOptions{})
+	if err != nil {
+		return nil, errors.Wrap(err, "error while getting storage class")
+	}
+	return sc, nil
+}
+
+// GetCSIControllerSTS returns the CSI controller sts with a specific
+// openebs-component-name label key
+func (k K8sClient) GetCSIControllerSTS(name string) (*appsv1.StatefulSet, error) {
+	if sts, err := k.K8sCS.AppsV1().StatefulSets("").List(context.TODO(), metav1.ListOptions{
+		LabelSelector: fmt.Sprintf("openebs.io/component-name=%s", name),
+	}); err == nil && len(sts.Items) == 1 {
+		return &sts.Items[0], nil
+	} else if sts != nil {
+		return nil, fmt.Errorf("got %d statefulsets with the label openebs.io/component-name=%s", len(sts.Items), name)
+	} else {
+		return nil, fmt.Errorf("got 0 statefulsets with the label openebs.io/component-name=%s", name)
+	}
+}
+
+// GetEvents returns the corev1 events based on the fieldSelectors
+func (k K8sClient) GetEvents(fieldSelector string) (*corev1.EventList, error) {
+	events, err := k.K8sCS.CoreV1().Events("").List(context.TODO(), metav1.ListOptions{FieldSelector: fieldSelector})
+	if err != nil {
+		return nil, fmt.Errorf("error getting events for the resource : %v", err)
+	}
+	return events, nil
+}
+
+/*
+	PERSISTENT VOLUMES AND CLAIMS
+*/
+
+// GetPV returns a PersistentVolume object using the pv name passed.
+func (k K8sClient) GetPV(name string) (*corev1.PersistentVolume, error) {
+	pv, err := k.K8sCS.CoreV1().PersistentVolumes().Get(context.TODO(), name, metav1.GetOptions{})
+	if err != nil {
+		return nil, errors.Wrap(err, "error while getting persistent volume")
+	}
+	return pv, nil
+}
+
+// GetPVs returns a list of PersistentVolumes based on the values of volNames slice.
+// volNames slice if is nil or empty, it returns all the PVs in the cluster.
+// volNames slice if is not nil or not empty, it return the PVs whose names are present in the slice.
+// labelselector takes the label(key+value) and makes an api call with this filter applied. Can be empty string if label filtering is not needed.
+func (k K8sClient) GetPVs(volNames []string, labelselector string) (*corev1.PersistentVolumeList, error) {
+	pvs, err := k.K8sCS.CoreV1().PersistentVolumes().List(context.TODO(), metav1.ListOptions{LabelSelector: labelselector})
+	if err != nil {
+		return nil, err
+	}
+	volMap := make(map[string]corev1.PersistentVolume)
+	for _, vol := range pvs.Items {
+		volMap[vol.Name] = vol
+	}
+	var list []corev1.PersistentVolume
+	if len(volNames) == 0 {
+		return pvs, nil
+	}
+	for _, name := range volNames {
+		if pool, ok := volMap[name]; ok {
+			list = append(list, pool)
+		} else {
+			fmt.Printf("Error from server (NotFound): PV %s not found\n", name)
+		}
+	}
+	return &corev1.PersistentVolumeList{
+		Items: list,
+	}, nil
+}
+
+// GetPVC returns a PersistentVolumeClaim object using the pvc name passed.
+func (k K8sClient) GetPVC(name string, namespace string) (*corev1.PersistentVolumeClaim, error) {
+	pvc, err := k.K8sCS.CoreV1().PersistentVolumeClaims(namespace).Get(context.TODO(), name, metav1.GetOptions{})
+	if err != nil {
+		return nil, errors.Wrap(err, "error while getting persistent volume claim")
+	}
+	return pvc, nil
+}
+
+// GetPVCs returns a list of PersistentVolumeClaims based on the values of pvcNames slice.
+// namespace takes the namespace in which PVCs are present.
+// pvcNames slice if is nil or empty, it returns all the PVCs in the cluster, in the namespace.
+// pvcNames slice if is not nil or not empty, it return the PVCs whose names are present in the slice, in the namespace.
+// labelselector takes the label(key+value) and makes an api call with this filter applied. Can be empty string if label filtering is not needed.
+func (k K8sClient) GetPVCs(namespace string, pvcNames []string, labelselector string) (*corev1.PersistentVolumeClaimList, error) {
+	pvcs, err := k.K8sCS.CoreV1().PersistentVolumeClaims(namespace).List(context.TODO(), metav1.ListOptions{LabelSelector: labelselector})
+	if err != nil {
+		return nil, err
+	}
+	if len(pvcNames) == 0 {
+		return pvcs, nil
+	}
+	pvcNamePVCmap := make(map[string]corev1.PersistentVolumeClaim)
+	for _, item := range pvcs.Items {
+		pvcNamePVCmap[item.Name] = item
+	}
+	var items = make([]corev1.PersistentVolumeClaim, 0)
+	for _, name := range pvcNames {
+		if _, ok := pvcNamePVCmap[name]; ok {
+			items = append(items, pvcNamePVCmap[name])
+		}
+	}
+	return &corev1.PersistentVolumeClaimList{
+		Items: items,
+	}, nil
 }
